Split frontend and backend
